Team detail screen working

+++ app/src/main/java/com/sgut/android/nationalfootballleague/data/remote/responses/teamdeats/Stats.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
 data class Stats (
 
   @SerializedName("name"  ) var name  : String? = null,
-  @SerializedName("value" ) var value : Int?    = null
+  @SerializedName("value" ) var value : Float?    = null
 
 )
